refactor [#260] 변경된 구조에 맞춰 아티스트 리졸버 수정
