+++ src/main/java/org/sopt/confeti/global/resolver/music_api/MusicAPISpecificResolver.java
@@ -1,0 +1,6 @@
+package org.sopt.confeti.global.resolver.music_api;
+
+public interface MusicAPISpecificResolver {
+
+    <T> void load(final T target);
+}
